refactor: 패키지 구조 변경 및 클래스 이름 변경
