+++ app/src/androidTest/java/woowacourse/shopping/view/cart/CartFragmentTest.kt
@@ -12,8 +12,8 @@ import org.hamcrest.Matchers.allOf
 import org.junit.Before
 import org.junit.Test
 import woowacourse.shopping.R
-import woowacourse.shopping.data.repository.CartRepository
 import woowacourse.shopping.data.repository.RepositoryProvider
+import woowacourse.shopping.data.repository.cart.CartRepository
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.fixture.dummyProductsFixture
 import woowacourse.shopping.presentation.view.cart.CartFragment
