+++ app/src/androidTest/java/woowacourse/shopping/view/catalog/CatalogFragmentTest.kt
@@ -14,8 +14,8 @@ import org.hamcrest.Matchers.allOf
 import org.junit.Before
 import org.junit.Test
 import woowacourse.shopping.R
-import woowacourse.shopping.data.repository.ProductRepository
 import woowacourse.shopping.data.repository.RepositoryProvider
+import woowacourse.shopping.data.repository.product.ProductRepository
 import woowacourse.shopping.domain.Product
 import woowacourse.shopping.fixture.dummyProductsFixture
 import woowacourse.shopping.presentation.view.catalog.CatalogFragment
