+++ app/src/main/java/woowacourse/shopping/data/datasource/coupon/CouponDataSource.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.datasource
+package woowacourse.shopping.data.datasource.coupon
 
 import woowacourse.shopping.domain.coupon.Coupon
