+++ app/src/main/java/woowacourse/shopping/data/datasource/order/OrderDataSource.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.datasource
+package woowacourse.shopping.data.datasource.order
 
 interface OrderDataSource {
     suspend fun submitOrder(cartIds: List<Long>)
