+++ app/src/main/java/woowacourse/shopping/data/datasource/product/ProductDataSource.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.datasource
+package woowacourse.shopping.data.datasource.product
 
 import woowacourse.shopping.domain.Page
 import woowacourse.shopping.domain.Product
