+++ app/src/main/java/woowacourse/shopping/data/repository/RepositoryProvider.kt
@@ -2,11 +2,19 @@ package woowacourse.shopping.data.repository
 
 import android.content.Context
 import woowacourse.shopping.data.api.ApiClient
-import woowacourse.shopping.data.datasource.CartItemDataSourceImpl
-import woowacourse.shopping.data.datasource.CouponDataSourceImpl
-import woowacourse.shopping.data.datasource.OrderDataSourceImpl
-import woowacourse.shopping.data.datasource.ProductDataSourceImpl
+import woowacourse.shopping.data.datasource.cart.CartDataSourceImpl
+import woowacourse.shopping.data.datasource.coupon.CouponDataSourceImpl
+import woowacourse.shopping.data.datasource.order.OrderDataSourceImpl
+import woowacourse.shopping.data.datasource.product.ProductDataSourceImpl
 import woowacourse.shopping.data.db.ShoppingDatabase
+import woowacourse.shopping.data.repository.cart.CartRepository
+import woowacourse.shopping.data.repository.cart.CartRepositoryImpl
+import woowacourse.shopping.data.repository.coupon.CouponRepository
+import woowacourse.shopping.data.repository.coupon.CouponRepositoryImpl
+import woowacourse.shopping.data.repository.order.OrderRepository
+import woowacourse.shopping.data.repository.order.OrderRepositoryImpl
+import woowacourse.shopping.data.repository.product.ProductRepository
+import woowacourse.shopping.data.repository.product.ProductRepositoryImpl
 
 object RepositoryProvider {
     private const val NOT_INITIALIZED_MESSAGE = "%s가 초기화되지 않았습니다"
@@ -56,7 +64,7 @@ object RepositoryProvider {
         val recentProductDao = database.recentProductDao()
 
         val productDataSource = ProductDataSourceImpl(ApiClient.productService)
-        val cartItemDataSource = CartItemDataSourceImpl(ApiClient.cartItemService)
+        val cartItemDataSource = CartDataSourceImpl(ApiClient.cartItemService)
         val couponDataSource = CouponDataSourceImpl(ApiClient.couponService)
         val orderDataSource = OrderDataSourceImpl(ApiClient.orderService)
