+++ app/src/main/java/woowacourse/shopping/data/repository/cart/CartRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository
+package woowacourse.shopping.data.repository.cart
 
 import woowacourse.shopping.domain.CartItem
 import woowacourse.shopping.domain.Page
