+++ app/src/main/java/woowacourse/shopping/data/repository/coupon/CouponRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository
+package woowacourse.shopping.data.repository.coupon
 
 import woowacourse.shopping.domain.coupon.Coupon
