+++ app/src/main/java/woowacourse/shopping/data/repository/order/OrderRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository
+package woowacourse.shopping.data.repository.order
 
 interface OrderRepository {
     suspend fun placeOrder(cartIds: List<Long>)
