+++ app/src/main/java/woowacourse/shopping/presentation/view/checkout/CheckoutEventHandler.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.presentation.view.checkout
 
-import woowacourse.shopping.presentation.view.checkout.adapter.CouponUiModel
+import woowacourse.shopping.presentation.model.CouponUiModel
 
 interface CheckoutEventHandler {
     fun onToggleSelection(
