+++ app/src/main/java/woowacourse/shopping/presentation/view/checkout/adapter/CouponViewHolder.kt
@@ -4,6 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCouponBinding
+import woowacourse.shopping.presentation.model.CouponUiModel
 import woowacourse.shopping.presentation.view.checkout.CheckoutEventHandler
 
 class CouponViewHolder(
